requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2requests-html==0.10.0google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
schedule==1.2.0
